requirements.txt
pandas==1.5.0
plotly==5.11.0
python_dateutil==2.8.2
streamlit==1.16.0
yfinance==0.2.3
